Autokey Scripts
- TerminalAK opens a prompt window for the terminal to prompt commands

-ClipOutput should be used after the assistant has output his answer. The scripts copies all messages under the last "2>/dev/null", which might need further improvement, to only copy output below the LAST "2>/dev/null"

I recommend "ctrl + m" for the terminal and "ctrl + w" for the copy. Use "ctrl + w", wait a second until a notification appears and then press "ctrl + shift + v" to copy the answer of the assistant. Note that you can change the shortcut to "ctrl + v" if you like... depending on the terminal type of course.